backend/requirements.txt
Django==3.2.8
djangorestframework==3.12.4
django-rest-auth==0.9.5
django-allauth==0.46.0
django-cors-headers==3.10.0
pip==21.3
setuptools==58.2.0
wheel==0.37.0
